+++ CHANGELOG.md
@@ -29,6 +29,24 @@ The general format is:
 
 ```
 
+# 2.0.0 - DATE (08/02/2022)
+
+### Added
+
+- support for custom resources on `Resource` definitions
+
+# 2.0.0rc1 - DATE (08/02/2022)
+
+### Added
+
+- support for multiple volumes (volumes, config maps, secrets, existing volumes)
+
+# 1.8.0rc7 - DATE (08/02/2022)
+
+### Added
+
+- add support for bucket resource inputs with key only
+
 # 1.8.0rc6 - DATE (08/02/2022)
 
 ### Changed